Implement storage locales and generalize stashes
CardStash is a general class representing a collection of cards. Deck is
a kind of CardStash with additional data on format and legality.
StorageLocale is a kind of CardStash with additional data on the storage
location of the stash and (in the future) pricing.

The cardstash folder contains the general templates and views shared by
CardStashes.